requirements.txt
numpy==1.24.3
pandas==2.0.3
aiohttp==3.8.5
pytz==2023.3
python-dotenv==1.0.0
# asyncio is part of the Python standard library, no need to install it separately
# Only pandas is needed for the news sources script
# Web requests and data processing
requests>=2.28.1
feedparser>=6.0.10
beautifulsoup4>=4.11.1
# New packages for ClickHouse and newswire monitoring
clickhouse-connect>=0.7.0
lxml>=4.9.3
# Brotli compression support for web scraping
brotli>=1.0.9
